Add deprecation warning if ImageGpu_*_C3 and stride() is used
In that case, the stride, computed by pitch / sizeof(PixelType),
can be a non-integer number. This yields wrong results. There is a
workaround for that when a pixel is accessed using the data() function
-> pitch is used instead of stride!

However, e.g. trust-functions use the stride function and they will not work
correct if pitch /sizeof(PixelType) is non-integer.
Additionally, kernels, which access directly the
stride function could be affected.

In either case, a runtime-warning is thrown if stride is non-integer.